examine_package: change handling of encodings and IO
This change should work in both python2 and python3.

Handle all data as str() in the functions that generate the data.

All functions in examine_package now return data, instead of writing it
themselves.

Don't replace stdout with the less fd, but write directly to less in
examine_package and process_new.

Force all output to less to be binary using six (which always uses
utf-8).

Dxplicitly open file as binary in strip_pgp_signature.

Write the html output to file as text in show_new.