dak rm: allow removal of binaries without source
When removing binaries from a source, do the removal, even if the source is
not in the suite.

Normally, if a binary is in the suite, the source of that binary should also
be in the suite. However, it seems that this is not always the case. As an
example, some binaries that were prop-uped into unstable don't seems to have
the corresponding source in unstable.

This change allow dak rm to remove these binaries.