generate-index-diffs: rewind `oldf` before using it
The `sizehashes(oldf)` call earlier changes the file position tot he
end of the file.  We need to rewind it before passing it to
subprocesses.

This was broken by removing the `pipe_file` function in
bad93415783955ad1d517ef53b31a2b3b003d03b which included the call to
`seek` before.