pdiff: Repair patch history when misaligned
Due to a previous bug in dak, the pdiff history in Debian experimental
is currently misaligned.  Unfortunately, dak is not self-healing with
this kind of history corruption (but it had been if the field order
had been different).

This patch will have dak detect misaligned patch histories (where
e.g. `SHA1-Patches` and `SHA256-Patches` disagree on the order) and
attempt to re-align them.  We re-align by looking from the end of the
two misaligned histories and keep all the patches they have in common.

This is done on the basis that we keep adding patches at the end, so
the history should be aligned there (assumping relevant bug fixes for
the corruption source have implemented).  Once we discover a
difference, we discard that and all older patches.  This is a
deliberate choice as our PDiff code assumes that the history is in
order without any gaps (plus it was easier that way).

In the worst case, this will reset the entire history and slow rebuild
from there.

Note that we repair unmerged and merged histories separately.  This
can cause one series to have more patches than the other.  This is
mostly a non-issue as dak will cope with it (as long as both series
are consistent internally).

Signed-off-by: Niels Thykier <niels@thykier.net>